+++ .gitignore
@@ -2,3 +2,4 @@ Sources/main.swift
 Packages
 .build
 Resources
+*.xcodeproj
