+++ Sources/MustacheRenderer.swift
@@ -1,12 +1,6 @@
 import Vapor
 import Mustache
 
-#if os(Linux)
-    import Glibc
-#else
-    import Darwin.C
-#endif
-
 public class MustacheRenderer: RenderDriver {
     static let currentName = "__current"
 
@@ -17,18 +11,10 @@ public class MustacheRenderer: RenderDriver {
 
         for (name, file) in files {
             do {
-                #if os(Linux)
-                    let bytes = try readBytesFromFile(path: file)
+                var bytes = try File.readBytes(path: file)
+                bytes.append(0)
 
-                    var signedData = bytes.map { byte in
-                        return Int8(byte)
-                    }
-
-                    signedData.append(0)
-                    includes[name] = String(validatingUTF8: signedData)
-                #else
-                    includes[name] = try String(contentsOfFile: file)
-                #endif
+                includes[name] = String(validatingUTF8: bytes)
             } catch {
                 Log.warning("Could not open file \(file). Error: \(error)")
             }
@@ -46,59 +32,4 @@ public class MustacheRenderer: RenderDriver {
         return result
     }
 
-    enum Error: ErrorProtocol {
-        case CouldNotOpenFile
-        case Unreadable
-    }
-
-    func readBytesFromFile(path: String) throws -> [UInt8] {
-        let fd = open(path, O_RDONLY);
-
-        if fd < 0 {
-            throw Error.CouldNotOpenFile
-        }
-        defer {
-            close(fd)
-        }
-
-        var info = stat()
-        let ret = withUnsafeMutablePointer(&info) { infoPointer -> Bool in
-            if fstat(fd, infoPointer) < 0 {
-                return false
-            }
-            return true
-        }
-        
-        if !ret {
-            throw Error.Unreadable
-        }
-        
-        let length = Int(info.st_size)
-        
-        let rawData = malloc(length)
-        var remaining = Int(info.st_size)
-        var total = 0
-        while remaining > 0 {
-            let advanced = rawData.advanced(by: total)
-            
-            let amt = read(fd, advanced, remaining)
-            if amt < 0 {
-                break
-            }
-            remaining -= amt
-            total += amt
-        }
-
-        if remaining != 0 {
-            throw Error.Unreadable
-        }
-
-        //thanks @Danappelxx
-        let data = UnsafeMutablePointer<UInt8>(rawData)
-        let buffer = UnsafeMutableBufferPointer<UInt8>(start: data, count: length)
-        return Array(buffer)
-    }
-
 }
-
-
